+++ packages/cli-tools/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support path notation when defining a stream ID
 - Replaced `stream list --operation` with `stream list --permission`
   - uses the same enum values as `stream grant-permission`
+- Removed `lastUpdated` and `name` columns from `stream list` outpuut
 - Modified configuration management:
   - can read config from `~/.streamr/config/default.json`
   - `--config` argument added